pyproject.toml
[project]
name = "2-llm-modelisation"
version = "0.1.0"
description = "Partie de modélisation LLM : configuration, personnalisation et intégration d’un modèle de langage large (LLM) afin de permettre l’analyse, l’indexation et le traitement avancé des documents collectés lors du scraping."
authors = ["Malek BOUMEDINE <malek.boumedine@gmail.com>"]
keywords = ["scraping", "data", "LLM", "RAG"]
licence = "MIT"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "alembic>=1.16.4",
    "chromadb>=1.0.15",
    "dotenv>=0.9.9",
    "fastapi>=0.116.1",
    "groq>=0.30.0",
    "jose>=1.0.0",
    "langchain-community>=0.3.27",
    "langchain-core>=0.3.68",
    "langchain-groq>=0.3.6",
    "langchain-huggingface>=0.3.0",
    "langchain-ollama>=0.3.4",
    "langchain-perplexity>=0.1.1",
    "langchain-qdrant>=0.2.0",
    "langchain-tavily>=0.2.8",
    "langgraph>=0.5.2",
    "networkx>=3.5",
    "ollama>=0.5.1",
    "openpyxl>=3.1.5",
    "pandas>=2.3.1",
    "passlib>=1.7.4",
    "pdfplumber>=0.11.7",
    "psycopg2-binary>=2.9.10",
    "pydantic[email]>=2.11.7",
    "pypdf>=5.7.0",
    "pypdf2>=3.0.1",
    "python-jose>=3.5.0",
    "python-multipart>=0.0.20",
    "qdrant-client[fastembed]>=1.14.3",
    "sentence-transformers>=5.0.0",
    "sqlmodel>=0.0.24",
    "tavily-python>=0.7.9",
    "uvicorn>=0.35.0",
    "xlrd>=2.0.2",
]
